chore: rename test to tests
